refactor(wow-openapi): extract CommandRequestParameters and CommandResponses
- Extract CommandRequestParameters into separate file
- Extract CommandResponses into separate file
- Update imports and usages in related files
- Add EmptyClassConverter for handling empty schemas
